infrastructure/cloud-functions/instatter-poc/requirements.txt
functions-framework>=3.5.0
google-cloud-storage>=2.16.0
playwright>=1.52.0
